add: Add MediaPlayer seek accuracy test
Seeking a video to a certain position may not result in seeking to that exact position.
This is because videos are seeking to the nearest previous keyframe position.
It seems like seeking to an approximately accurate position will be supported in Android O.
Constant integers are reserved for setting seek mode.